On x86, load initrd using DMA interface.
This makes it infinitely faster.

Also, I changed pc-bios/linuxboot.bin from a checked in binary to a
symlink, since that seems to be what the build uses. (Otherwise, it'll
build a linuxboot.bin in the optionrom subdirectory, but helpfully not
use it.) Possibly there's a better way to do this.

Change-Id: Ic3373c2b03e40b9cf461efc848bca8c39c7a312d